Facade is mostly (done?) and somewhat tested.
I need to go higher level tomorrow and think how I'll implement a save feature, so that the app can stay on and we can simply press on/off and it'll remember the websites it wants blocked. I'm thinking this needs to be at a higher level where we just request the active urls from the facade, store them, unblock everything when turned off, then when turned back on go into storage and send all of those to the facade.

This way, unblocking a website automatically deletes it but turning the feature off does not. When you want to get back into focus, simply turn it back on and your list is there. Imagine how ANNOYING it would be to have to insert all your websites one by one again and again everytime you want to use the app.